requirements.txt
streamlit
openai-whisper
gTTS
